Dear Professor Pilcher and the Graders,
Attached, please find my first midterm exam for the Mathematical and Statistical Computing Class.  There are 2 files: 1 for the Python and the other for the R-Studio Problem.  I worked very hard on it.  Please let me know if there is any problem.  Thanking you.

Also, for the 2nd problem, although Sean had put the value 53490 in the Blackboard notes for guidance, I found that it may be a typo because the value returned in my R-Studio was 53,940.  It was a slight discrepancy, but I stuck with my original answer.  Thanking you again! I wanted to humbly point that out.

Best wishes,

Saniya Khullar